added confirm delete logics

+++ src/app/components/invoice-detail-header/invoice-detail-header.component.ts
@@ -11,16 +11,23 @@ import {
 import { InvoiceFormComponent } from '../invoice-form/invoice-form.component';
 import { SidebarModule } from 'primeng/sidebar';
 import { ModalService } from '../../services/modalService/modal.service';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ToastModule } from 'primeng/toast';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
 
 @Component({
   selector: 'app-invoice-detail-header',
   standalone: true,
-  imports: [StatusComponent, InvoiceFormComponent, SidebarModule, ToastModule],
+  imports: [
+    StatusComponent,
+    InvoiceFormComponent,
+    SidebarModule,
+    ToastModule,
+    ConfirmDialogModule,
+  ],
   templateUrl: './invoice-detail-header.component.html',
   styleUrl: './invoice-detail-header.component.sass',
-  providers: [MessageService],
+  providers: [MessageService, ConfirmationService],
 })
 export class InvoiceDetailHeaderComponent {
   @Input() invoice!: Invoice;
@@ -30,17 +37,42 @@ export class InvoiceDetailHeaderComponent {
     private router: Router,
     private store: Store<AppState>,
     private modalService: ModalService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ) {
     this.modalService.isModalOpen('editInvoice').subscribe((isOpen) => {
       this.isEditModalOpen = isOpen;
     });
   }
 
-  deleteInvoice = (id: string) => {
-    this.store.dispatch(deleteInvoice({ id }));
-    this.router.navigate(['']);
-  };
+  confirmDelete(id: string) {
+    this.confirmationService.confirm({
+      header: 'Confirmation Deletion',
+      message: `Are you sure you want to delete invoice # ${id}? This action cannot be undone.`,
+      acceptIcon: 'pi pi-check mr-2',
+      rejectIcon: 'pi pi-times mr-2',
+      rejectButtonStyleClass: 'btn btn-secondary',
+      acceptButtonStyleClass: 'btn btn-warning',
+      accept: () => {
+        this.store.dispatch(deleteInvoice({ id }));
+        this.messageService.add({
+          severity: 'success',
+          summary: 'Delete Item',
+          detail: `Item ${id} Deleted`,
+          life: 10000,
+        });
+        this.router.navigate(['']);
+      },
+      // reject: () => {
+      //   this.messageService.add({
+      //     severity: 'error',
+      //     summary: 'Rejected',
+      //     detail: 'You have rejected',
+      //     life: 2000,
+      //   });
+      // },
+    });
+  }
 
   markAsPaid = (id: string) => {
     this.store.dispatch(updateInvoice({ invoice: { id, status: 'paid' } }));
